Fix max_input_time
762fd8c0c06a143a49efc006da93caa49c975e00 turned off nginx's request
buffering, causing the input from the brwoser to be feed
to PHP straight away, without the nginx webserver saving it to a
temporary file first.

Raise PHP's max_input_time to prevent timeouts (resulting in a 502
"Bad gateway" page being served to the user) if the upload takes
too long.