Add support for provisioning CM3(+) modules
On CM3(+) it uses USB instead of Ethernet for data transfer.

Requires that server is using predictable (based on MAC-address)
link-local IPv6 addresses.
If server is using RPI OS dhcpcd configuration will be changed
automatically by the .deb for to accomodate this.
If you are running any other Linux distribution on the server,
it is up to you to configure this.