Improve error reporting when image does not fit in storage
Ref #15

Report an error if image is larger than capacity of eMMC/SD card.
Also report it if /dev/mmcblk0 is missing (e.g. CM4 Lite without
card).

(The check does require that cmprovision knows the uncompressed
image size of the image.
It takes a some minutes after uploading a new image
before this is known, as it computes the uncompressed size
and SHA256 in the background.
If user attempts to provision before that the check is skipped.)